+++ feature-voting/src/main/java/co/electriccoin/zcash/ui/screen/voting/polldescription/VotePollDescriptionView.kt
@@ -59,7 +59,7 @@ fun VotePollDescriptionView(
                 )
                 Surface(
                     shape = CircleShape,
-                    color = ZashiColors.Surfaces.bgSecondary,
+                    color = ZashiColors.Surfaces.bgPrimary,
                     modifier =
                         Modifier
                             .align(Alignment.CenterStart)